+++ Core/Blacklist.lua
@@ -39,6 +39,9 @@ function Blacklist:Initialize()
             if not SimpleDisenchantDB.blacklist then
                 SimpleDisenchantDB.blacklist = {}
             end
+            if not SimpleDisenchantDB.minimap then
+                SimpleDisenchantDB.minimap = { angle = 225, hidden = false }
+            end
             blacklistedItems = SimpleDisenchantDB.blacklist
             initialized = true
             self:UnregisterEvent("ADDON_LOADED")